requirements.txt
asgiref==3.2.9
Babel==2.8.0
boto3==1.14.4
botocore==1.17.4
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-phonenumber-field==4.0.0
django-resized==0.3.11
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
idna==2.9
jmespath==0.10.0
oauthlib==3.1.0
phonenumbers==8.12.5
Pillow==7.1.2
psycopg2==2.7.6
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
twilio==6.42.0
urllib3==1.25.9
whitenoise==5.0.1
